requirements.txt
annotated-types==0.7.0
anyio==4.10.0
click==8.2.1
colorama==0.4.6
exceptiongroup==1.3.0
fastapi==0.110.1
h11==0.16.0
idna==3.10
numpy==2.2.6
pandas==2.2.2
pydantic==2.11.7
pydantic_core==2.33.2
pyodbc==5.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2025.2
six==1.17.0
sniffio==1.3.1
starlette==0.37.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uvicorn[standard]==0.29.0
openpyxl>=3.1.2
sqlalchemy==2.0.30